Generation et stockage local des trajets. Save dans WIP, requiert lecture correcte json

+++ server/src/main/java/com/horarbus/service/MapsService.java
@@ -25,10 +25,10 @@ public class MapsService {
     }
 
     public static String getItinerary(String startingPlaceId, String arrivalPlaceId,
-            TravelMode travelMode, int arrivalTime) throws Exception {
+            TravelMode travelMode, long arrivalTime) throws Exception {
         String params = "origin=place_id:" + formatValue(startingPlaceId) + "&destination=place_id:"
                 + formatValue(arrivalPlaceId) + "&mode=" + travelMode.toUrlValue()
-                + "&arrival_time=" + Integer.toString(arrivalTime);
+                + "&arrival_time=" + Long.toString(arrivalTime);
 
         String endpointStr = MapsService.ITINERARY_URL + params;
         return MapsService.doRequest(generateEndpoint(endpointStr));
